feat: use plugins effects on fetchNextBatch listener
