+++ plugins-structure/packages/ticketvote/src/effects/index.js
@@ -1,0 +1,30 @@
+import { ticketvoteSummaries } from "../ticketvote/summaries";
+import { getTokensToFetch } from "@politeiagui/core/records/utils";
+import isEmpty from "lodash/isEmpty";
+
+export async function fetchNextTicketvoteSummaries(
+  state,
+  dispatch,
+  { inventoryList }
+) {
+  const {
+    ticketvoteSummaries: { byToken, status },
+    ticketvotePolicy: {
+      policy: { summariespagesize },
+    },
+  } = state;
+
+  const voteSummariesToFetch = getTokensToFetch({
+    inventoryList,
+    lookupTable: byToken,
+    pageSize: summariespagesize,
+  });
+
+  if (status !== "loading" && !isEmpty(voteSummariesToFetch)) {
+    await dispatch(
+      ticketvoteSummaries.fetch({
+        tokens: voteSummariesToFetch,
+      })
+    );
+  }
+}
